finish part 1 of output test

+++ Assignment2/point.cpp
@@ -5,18 +5,18 @@ using namespace std;
 Point::Point()
 	:x(0), y(0)
 {}
-Point::Point(int initialX, int initialY)
+Point::Point(double initialX, double initialY)
 	:x(initialX), y(initialY)
 {}
-int Point::getX() const
+double Point::getX() const
 {
 	return x; 
 }
-int Point::getY() const
+double Point::getY() const
 {
 	return y;
 }
-void Point::setPoint(int newX, int newY)
+void Point::setPoint(double newX, double newY)
 {
 	x = newX;
 	y = newY;
